requirements.txt
django==1.11
django-allauth==0.31.0
django-extensions==1.7.8
mysqlclient==1.3.10
